+++ crates/ruma-signatures/src/ed25519/compat.rs
